+++ R/cox_regression.R
@@ -110,11 +110,11 @@ h_coxreg_univar_formulas <- function(variables,
   }
 
   assertthat::assert_that(
-    is_variables(variables[c(arm_name, "event", "time")]),
     assertthat::is.flag(interaction),
     (has_arm || (!interaction)),
     (!is.null(variables$covariates) || (!interaction))
   )
+  assert_list_of_variables(variables[c(arm_name, "event", "time")])
 
   if (!is.null(variables$covariates)) {
     forms <- paste0(
@@ -197,9 +197,7 @@ h_coxreg_multivar_formula <- function(variables) {
     assertthat::assert_that(is.character(variables$covariates))
   }
 
-  assertthat::assert_that(
-    is_variables(variables[c(arm_name, "event", "time")])
-  )
+  assert_list_of_variables(variables[c(arm_name, "event", "time")])
 
   y <- paste0(
     "survival::Surv(", variables$time, ", ", variables$event, ") ~ ",
@@ -332,9 +330,10 @@ fit_coxreg_univar <- function(variables,
   }
 
   assertthat::assert_that(
-    is_variables(variables[c(arm_name, "event", "time")]),
     is_df_with_variables(data, as.list(unlist(variables)))
   )
+  assert_list_of_variables(variables[c(arm_name, "event", "time")])
+
   if (!is.null(variables$strata)) {
     assertthat::assert_that(
       control$pval_method != "likelihood"
@@ -606,10 +605,10 @@ fit_coxreg_multivar <- function(variables,
   }
 
   assertthat::assert_that(
-    is_variables(variables[c(arm_name, "event", "time")]),
     is_df_with_variables(data, as.list(unlist(variables))),
     isFALSE(control$interaction)
   )
+  assert_list_of_variables(variables[c(arm_name, "event", "time")])
 
   if (!is.null(variables$strata)) {
     assertthat::assert_that(