+++ tests/testthat/test-utils_checkmate.R
@@ -27,19 +27,19 @@ testthat::test_that("assert_nonnegative_count is FALSE with wrong input", {
   testthat::expect_error(assert_nonnegative_count(NA_integer_))
 })
 
-# is_variables ----
+# assert_list_of_variables ----
 
-testthat::test_that("is_variables is TRUE with healthy input", {
-  testthat::expect_true(is_variables(list(a = "bla", b = "bli")))
-  testthat::expect_true(is_variables(list(a = "123")))
-  testthat::expect_true(is_variables(list(a = c("bla", "bli"))))
+testthat::test_that("assert_list_of_variables is TRUE with healthy input", {
+  testthat::expect_silent(assert_list_of_variables(list(a = "bla", b = "bli")))
+  testthat::expect_silent(assert_list_of_variables(list(a = "123")))
+  testthat::expect_silent(assert_list_of_variables(list(a = c("bla", "bli"))))
 })
 
-testthat::test_that("is_variables is FALSE with wrong input", {
-  testthat::expect_false(is_variables(list("bla", b = "bli")))
-  testthat::expect_false(is_variables(list(a = 1, b = "bla")))
-  testthat::expect_false(is_variables(c(a = "blo", b = "bla")))
-  testthat::expect_false(is_variables(c(a = 1, a = 2)))
+testthat::test_that("assert_list_of_variables is FALSE with wrong input", {
+  testthat::expect_error(assert_list_of_variables(list("bla", b = "bli")))
+  testthat::expect_error(assert_list_of_variables(list(a = 1, b = "bla")))
+  testthat::expect_error(assert_list_of_variables(c(a = "blo", b = "bla")))
+  testthat::expect_error(assert_list_of_variables(c(a = 1, a = 2)))
 })
 
 # is_df_with_variables ----