+++ NEWS.md
@@ -5,6 +5,7 @@
   `checkmate::assertMultiClass()` and made internal
 * `is_nonnegative_count` is now `assert_nonnegative_count` and uses 
   `checkmate::assert_integerish()` and made internal
+* `assert_list_of_variables` instead of `is_variables`
 * renamed `assertthat.R` into a more appropriate `utils_checkmate.R`
 * renamed `test-assertthat.R` into `test-utils_checkmate.R`